refactor: ajusta controllers de usuário, playlist e vídeos; corrige model e rotas de usuário

+++ controllers/videoController.js
@@ -1,8 +1,16 @@
 const Video = require("../models/video");
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
+const JWT_SECRET = process.env.JWT_SECRET;
 
 module.exports = {
   async inserir(req, res, next) {
     try {
+      const token = req.cookies.access_token;
+      const decoded = jwt.verify(token, JWT_SECRET);
+      const idUsuario = decoded.id;
+      req.body.usuarioId = idUsuario;
+
       const novoVideo = await Video.create(req.body);
       res.status(201).json(novoVideo);
     } catch (err) {
@@ -41,12 +49,29 @@ module.exports = {
 
   async atualizar(req, res, next) {
     try {
-      const video = await Video.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-      });
+      const token = req.cookies.access_token;
+      const decoded = jwt.verify(token, JWT_SECRET);
+
+      const idUsuario = decoded.id;
+
+      const video = await Video.findById(req.params.id);
+
       if (!video)
         return res.status(404).json({ error: "Vídeo não encontrado." });
-      res.json(video);
+
+      if (!video.usuarioId.equals(idUsuario)) {
+        const err = new Error(
+          "Você não tem permissão para atualizar este vídeo."
+        );
+        err.status = 403;
+        return next(err);
+      }
+      const videoAtualizado = await Video.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        { new: true }
+      );
+      res.json(videoAtualizado);
     } catch (err) {
       err.status = 400;
       err.message = "Erro ao atualizar vídeo.";
@@ -56,9 +81,26 @@ module.exports = {
 
   async deletar(req, res, next) {
     try {
-      const video = await Video.findByIdAndDelete(req.params.id);
+      const token = req.cookies.access_token;
+      const decoded = jwt.verify(token, JWT_SECRET);
+
+      const idUsuario = decoded.id;
+
+      const video = await Video.findById(req.params.id);
+
       if (!video)
         return res.status(404).json({ error: "Vídeo não encontrado." });
+
+      if (!video.usuarioId.equals(idUsuario)) {
+        const err = new Error(
+          "Você não tem permissão para atualizar este vídeo."
+        );
+        err.status = 403;
+        return next(err);
+      }
+
+      await Video.findByIdAndDelete(req.params.id);
+
       res.status(204).send();
     } catch (err) {
       err.status = 400;
